fix build

+++ pkg/ddl/index.go
@@ -1737,7 +1737,7 @@ func (w *worker) onCreateFulltextIndex(jobCtx *jobContext, job *model.Job) (ver 
 			if done, err = tici.CheckAddIndexProgress(jobCtx.stepCtx, jobCtx.store, tblInfo.ID, indexInfo.ID); err != nil {
 				logutil.DDLLogger().Warn("[ddl] check TiCI fulltext index progress failed",
 					zap.Int64("jobID", job.ID),
-					zap.Int8("analyzeState", int8(job.ReorgMeta.AnalyzeState)),
+					zap.Int8("analyzeState", job.ReorgMeta.AnalyzeState),
 					zap.Int64("tableID", tblInfo.ID),
 					zap.Int64("indexID", indexInfo.ID),
 					zap.Error(err))
@@ -1746,7 +1746,7 @@ func (w *worker) onCreateFulltextIndex(jobCtx *jobContext, job *model.Job) (ver 
 			if !done {
 				logutil.DDLLogger().Debug("[ddl] wait for TiCI fulltext index ready",
 					zap.Int64("jobID", job.ID),
-					zap.Int8("analyzeState", int8(job.ReorgMeta.AnalyzeState)),
+					zap.Int8("analyzeState", job.ReorgMeta.AnalyzeState),
 					zap.Int64("tableID", tblInfo.ID),
 					zap.Int64("indexID", indexInfo.ID))
 				if err := waitTiCIAddIndexProgressPoll(jobCtx.stepCtx); err != nil {
@@ -1925,7 +1925,7 @@ func (w *worker) onCreateHybridIndex(jobCtx *jobContext, job *model.Job) (ver in
 			if done, err = tici.CheckAddIndexProgress(jobCtx.stepCtx, jobCtx.store, tblInfo.ID, indexInfo.ID); err != nil {
 				logutil.DDLLogger().Warn("[ddl] check TiCI hybrid index progress failed",
 					zap.Int64("jobID", job.ID),
-					zap.Int8("analyzeState", int8(job.ReorgMeta.AnalyzeState)),
+					zap.Int8("analyzeState", job.ReorgMeta.AnalyzeState),
 					zap.Int64("tableID", tblInfo.ID),
 					zap.Int64("indexID", indexInfo.ID),
 					zap.Error(err))
@@ -1934,7 +1934,7 @@ func (w *worker) onCreateHybridIndex(jobCtx *jobContext, job *model.Job) (ver in
 			if !done {
 				logutil.DDLLogger().Debug("[ddl] wait for TiCI hybrid index ready",
 					zap.Int64("jobID", job.ID),
-					zap.Int8("analyzeState", int8(job.ReorgMeta.AnalyzeState)),
+					zap.Int8("analyzeState", job.ReorgMeta.AnalyzeState),
 					zap.Int64("tableID", tblInfo.ID),
 					zap.Int64("indexID", indexInfo.ID))
 				if err := waitTiCIAddIndexProgressPoll(jobCtx.stepCtx); err != nil {
